+++ lib/src/framework/data_type/method_bindings/page_controller/register_fn.dart
@@ -1,0 +1,12 @@
+import '../../adapted_types/page_controller.dart';
+import '../method_binding_registry.dart';
+import 'commands.dart';
+
+void registerMethodCommandsForPageController(MethodBindingRegistry registry) {
+  registry.registerMethods<AdaptedPageController>({
+    'jumpToPage': PageControllerJumpToPageCommand(),
+    'animateToPage': PageControllerAnimateToPageCommand(),
+    'nextPage': PageControllerNextPageCommand(),
+    'previousPage': PageControllerPreviousPageCommand(),
+  });
+}