+++ lib/src/framework/data_type/method_bindings/register_bindings.dart
@@ -1,6 +1,8 @@
 import 'api_cancel_token/register_fn.dart';
 import 'async_controller/register_fn.dart';
+import 'file/register_fn.dart';
 import 'method_binding_registry.dart';
+import 'page_controller/register_fn.dart';
 import 'scroll_controller/register_fn.dart';
 import 'stream_controller/register_fn.dart';
 import 'text_field_controller/register_fn.dart';
@@ -13,4 +15,6 @@ void registerBindings(MethodBindingRegistry registry) {
   registerMethodCommandsForAsyncController(registry);
   registerMethodCommandsForTextFieldController(registry);
   registerMethodCommandsForApiCancelToken(registry);
+  registerMethodCommandsForFile(registry);
+  registerMethodCommandsForPageController(registry);
 }